finished day 33 by little tkinter app with quotes, changed the name of the main file to main_1 where i am trying to catch the iss when it  s above the head
